pyproject.toml
[project]
name = "benchmark-pipeline-dev-v4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anomalib>=2.2.0",
    "fiftyone>=1.11.0",
    "huggingface-hub>=1.3.1",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "scipy>=1.16.3",
    "tabulate>=0.9.0",
    "tifffile>=2025.12.20",
    "timm>=1.0.24",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
